Optimization

+++ src/main/java/ca/bcit/abalone/ai/Zobrist.java
@@ -6,6 +6,6 @@ public abstract class Zobrist<P, S, A, G extends Game<P, S, A>> {
 
     protected long[][] zobrist;
 
-    abstract public int hashCode(G game);
+    abstract public long hashCode(G game);
 
 }
